+++ app/src/main/java/com/tomaszrykala/wordduel/main/ui/SoftKeyboard.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.tomaszrykala.wordduel.game.keyboard.KeyTile
-import com.tomaszrykala.wordduel.game.state.KeyTiles
+import com.tomaszrykala.wordduel.game.board.KeyTiles
 
 @Composable
 fun SoftKeyboard(
